Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpParameters-degradationPreference-expected.txt


FAIL setParameters with degradationPreference set should succeed assert_equals: Expect dictionary.rtcp to be plain object expected "object" but got "undefined"
FAIL setParameters with degradationPreference unset should succeed assert_equals: Expect dictionary.rtcp to be plain object expected "object" but got "undefined"
